==> R/3_hvg_DEGs_HSC_scnym_annotation.Rmd <==
---
title: "scNym clustering HSC DEG"
output: html_notebook
---

```{r}
knitr::opts_chunk$set(comment = '', fig.width = 6, 
                      fig.height = 6, echo = FALSE, 
                      warning = FALSE)
```


```{r, message = FALSE}
library(miloR)
library(SingleCellExperiment)
library(scater)
library(dplyr)
library(patchwork)
library(here)
library(scales)
library(scran)
library(tidyverse)
library(ggrepel)
library(rcartocolor)
library(circlize)
library(Seurat)

outdir = here("Rout")

source(here("R/functions.R"))

library(loomR)
```

### Getting scnym-based leiden mapping 

```{r}
l1_file = here("scfiles/SA_HSC_v2_2608.loom")
lfile <- connect(filename = l1_file, mode = "r+", skip.validate = TRUE)

cattrnames = names(lfile$col.attrs)

SA.HSC.sample.annots = data.frame(leiden = lfile$col.attrs$leiden[])
for(attname in cattrnames) {
    SA.HSC.sample.annots[[attname]] = lfile$col.attrs[[attname]][]
}
lfile$close_all()
```

### Reading the hvg data

```{r}
l1_file = here("scfiles/hvg_HSC_X_scVI_based_200924.loom")
lfile <- connect(filename = l1_file, mode = "r+", skip.validate = TRUE)

cattrnames = names(lfile$col.attrs)

hvg.HSC.sample.annots = data.frame(leiden = lfile$col.attrs[["leiden"]][])
for(attname in cattrnames) {
    hvg.HSC.sample.annots[[attname]] = lfile$col.attrs[[attname]][]
}

hvg.HSC.v2 = SingleCellExperiment(list(logcounts=t(lfile[["matrix"]][,])),
    colData=hvg.HSC.sample.annots,
    rowData=DataFrame(gene_names=lfile$row.attrs$var_names[]))

# hvg.subset.HSC = hvg.subset.HSC[, colSums(hvg.subset.HSC@assays@data$logcounts) > 50]
# hvg.subset.HSC = hvg.subset.HSC[rowSums(hvg.subset.HSC@assays@data$logcounts) > 50,]

hvg.HSC.v2@assays@data$counts = exp(hvg.HSC.v2@assays@data$logcounts) - 1

rownames(hvg.HSC.v2) = rowData(hvg.HSC.v2)$gene_names
colnames(hvg.HSC.v2) = hvg.HSC.v2$obs_names

hvg.HSC.v2 = as.Seurat(hvg.HSC.v2)

lfile$close_all()
```

### Reading the raw data

```{r}
bm.final = readRDS(here("data/rds_files", "bm_final.rds") )
bm.final$batch = "final"
bm.pilot = readRDS(here("data/rds_files", "bm_pilot.rds") )
bm.pilot$batch = "pilot"

bm.merged.allgenes = merge(bm.final, bm.pilot, project = "bm.merged")


counts <- GetAssayData(bm.merged, layer="counts", assay="RNA")   
common.genes = intersect(Features(bm.merged.allgenes), Features(hvg.HSC.v2))

counts.sub <- counts[common.genes,]
bm.merged <- CreateSeuratObject(counts=counts.sub)
bm.merged@meta.data = bm.merged.allgenes@meta.data

```



### Running cluster DEGs

```{r}
# lfc.limit = 0.5

clust.HSC.DE.list = list()
hvg.HSC.sample.annots$leiden = as.character(hvg.HSC.sample.annots$leiden)
for (clust.id in unique(hvg.HSC.sample.annots$leiden)) {
    try({
        cat("cluster", clust.id, "\n")
    
    common.cells = intersect(rownames(bm.merged@meta.data),
                                             hvg.HSC.sample.annots$obs_names)
    
    cluster.raw.counts = bm.merged[rownames(bm.merged), 
                                   common.cells]
    
    bm.merged.metadata = bm.merged@meta.data[common.cells, ]
    
    rownames(hvg.HSC.sample.annots) = hvg.HSC.sample.annots$obs_names
    bm.merged.metadata$leiden = hvg.HSC.sample.annots[common.cells,]$leiden
    
    bm.merged.metadata = bm.merged.metadata %>% 
        mutate(cluster.condition = ifelse(leiden == clust.id, "1", "0"),
               cluster.condition = factor(cluster.condition, levels = c("0", "1")))
    
    cluster.subset.deseq = DESeqDataSetFromMatrix(cluster.raw.counts@assays$RNA$counts + 1,
                                                   colData = bm.merged.metadata,
                                                   design = ~ cluster.condition + batch)
    cluster.DE = pseudobulk_DEseq2(cluster.subset.deseq, contrast.in = c("cluster.condition", "1", "0"))

        # rld = rlog(cluster.subset.deseq, blind = FALSE)
        # assay(rld) <- limma::removeBatchEffect(assay(rld), rld$batch)

    cluster.DE = cluster.DE %>% arrange(padj, desc(log2FoldChange ) ) 
    clust.HSC.DE.list[[clust.id]] = cluster.DE
    
    cluster.DE %>% rename(names = gene) %>% rownames_to_column() %>%
        openxlsx::write.xlsx(file = here(paste0("DEG_sheets/cluster_markers/HSC/cluster", clust.id, ".HSC_X_scVI.20.09.2024.xlsx")))

    add_mouse_geneinfo(here(paste0("DEG_sheets/cluster_markers/HSC/cluster", clust.id,
                                   ".HSC_X_scVI.20.09.2024.xlsx")))
    })
}
saveRDS(clust.HSC.DE.list, file = here("data/rds_files/HSC.clusterDEGs.20092024.RDS") )
```


### Running DEG PBS vs SA

```{r}
lfc.limit = 0.5

clust.HSC.DE.list = list()
hvg.HSC.sample.annots$leiden = as.character(hvg.HSC.sample.annots$leiden)
for (clust.id in unique(hvg.HSC.sample.annots$leiden)) {
    try({
        cat("cluster", clust.id, "\n")
    
    hvg.clust.id = hvg.HSC.sample.annots %>% filter(leiden == clust.id)
    
    cluster.raw.counts = bm.merged[rownames(bm.merged), 
                                   intersect(rownames(bm.merged@meta.data),
                                             hvg.clust.id$obs_names)]
    
    bm.merged.metadata = bm.merged@meta.data %>% 
        mutate(condition = gsub("HTO-([A-Z]+)[0-9]", "\\1", sample))
    
    bm.merged.metadata = bm.merged.metadata[colnames(cluster.raw.counts), ] 
    
    cluster.subset.deseq = DESeqDataSetFromMatrix(cluster.raw.counts@assays$RNA@counts + 1,
                                                   colData = bm.merged.metadata,
                                                   design = ~ condition + batch)
    cluster.DE = pseudobulk_DEseq2(cluster.subset.deseq)

        # rld = rlog(cluster.subset.deseq, blind = FALSE)
        # assay(rld) <- limma::removeBatchEffect(assay(rld), rld$batch)

    clust.HSC.DE.list[[clust.id]] = cluster.DE
    # MA.cluster.normed.data[[clust.id]] =  rld
    
    cluster.DE = cluster.DE %>% filter(padj < 0.05, abs(log2FoldChange) > lfc.limit )
    clust.HSC.DE.list[[clust.id]] = cluster.DE
    
    cluster.DE %>% rename(names = gene) %>% rownames_to_column() %>%
        openxlsx::write.xlsx(file = here(paste0("DEG_sheets/cluster", clust.id, ".HSC.DEG.SAvsPBS.09.09.2024.xlsx")))

    add_mouse_geneinfo(here(paste0("DEG_sheets/cluster", 
                                   clust.id, ".HSC.DEG.SAvsPBS.09.09.2024.xlsx")))
    })
}
```

## Getting module scores

I decided to calculate the module scores for the batches separately, to not 
counfound the modulescores by batch effects.

### Module scores for EOS paper Jorssen et al 2024 Immunity Table 2

#### Reading table 2

```{r}
table2.files.dir = here("eos_maturation_paper_data/table_S2_markers/")

cell.type.genes = list()

for (csv.file in list.files(path = table2.files.dir, pattern = "*.csv", full.names = FALSE)) {
    cell.type = gsub(".csv", "", csv.file)
    cat(cell.type, "\n")
    cell.type.genes[[cell.type]] = read.table(
        file.path(table2.files.dir, csv.file),
        sep = "\t", h = T)
}

table2.gene.features = sapply(cell.type.genes, function(x) x$Gene)
```

#### Reading table 5

```{r}
table5.files.dir = here("eos_maturation_paper_data/table_S5_markers/")

cell.type.genes = list()

for (csv.file in list.files(path = table5.files.dir, pattern = "*.csv", full.names = FALSE)) {
    cell.type = gsub(".csv", "", csv.file)
    cat(cell.type, "\n")
    cell.type.genes[[cell.type]] = read.table(
        file.path(table5.files.dir, csv.file),
        sep = "\t", h = T)
}

table5.gene.features = sapply(cell.type.genes, function(x) x$Gene)
```

#### Reading gene features extracted by the Quant-seq experiment

```{r}
quantseq.DEGs = read_excel_allsheets(here("../SA_Quantseq2/outputs/DE.genes.SAvsPBS.xlsx"))
gene.score.features = sapply(quantseq.DEGs, function(x) x %>% filter(padj < 0.05) %>% pull(gene_name))
```

#### Reading cytopus gene sets 

```{r}
cytopus.dir = here("cytopus_cell_types")
cytopus.gene.sets = list()
for (filename in list.files(cytopus.dir)) {
    # cellname = gsub(".dat", "", filename)
    cellname = filename
    cytopus.gene.sets[[cellname]] = readLines(file.path(cytopus.dir, filename))
}
```


#### Reading panglodb gene sets 

```{r}
panglodb.dir = here("panglodb_dataset/pangledb_celltypes")
panglodb.gene.sets = list()
for (filename in list.files(panglodb.dir)) {
    # cellname = gsub(".dat", "", filename)
    cellname = filename
    panglodb.gene.sets[[cellname]] = readLines(file.path(panglodb.dir, filename))
}
```


Calculating the modulescores for pilot and final experiments:

```{r}
bm.pilot = AddModuleScore(bm.pilot, table2.gene.features, 
                                name = names(table2.gene.features))
bm.final = AddModuleScore(bm.final, table2.gene.features, 
                                name = names(table2.gene.features))


bm.pilot = AddModuleScore(bm.pilot, table5.gene.features, 
                                name = names(table5.gene.features))
bm.final = AddModuleScore(bm.final, table5.gene.features, 
                                name = names(table5.gene.features))


bm.pilot = AddModuleScore(bm.pilot, gene.score.features, 
                                name = names(gene.score.features))
bm.final = AddModuleScore(bm.final, gene.score.features, 
                                name = names(gene.score.features))


bm.pilot = AddModuleScore(bm.pilot, cytopus.gene.sets, 
                                name = names(cytopus.gene.sets))
bm.final = AddModuleScore(bm.final, cytopus.gene.sets, 
                                name = names(cytopus.gene.sets))
```


### hvg modulescores

```{r}
hvg.HSC.v2 = AddModuleScore(hvg.HSC.v2, table2.gene.features, 
                                name = names(table2.gene.features))

hvg.HSC.v2 = AddModuleScore(hvg.HSC.v2, table5.gene.features, 
                                name = names(table5.gene.features))

hvg.HSC.v2 = AddModuleScore(hvg.HSC.v2, gene.score.features, 
                                name = names(gene.score.features))

cytopus.gene.sets$FDC.dat = NULL
hvg.HSC.v2 = AddModuleScore(hvg.HSC.v2, cytopus.gene.sets, 
                                name = names(cytopus.gene.sets))

hvg.HSC.v2 = AddModuleScore(hvg.HSC.v2, panglodb.gene.sets, 
                                name = names(panglodb.gene.sets))
```

### Merging the bm final and bm pilot scores

```{r}
bm.modulescore.merged = merge(bm.pilot, bm.final, project = "modulscoring")
SA.HSC.modulescore.bm = bm.modulescore.merged[, SA.HSC.sample.annots$obs_names ]
```

### Writing the modulescores for quantseq gene sets

```{r}
quantseq.modulescores = SA.HSC.modulescore.bm@meta.data[, c("Eosinophil1" , 
                                       "Neutrophil2" , 
                                       "Neutrophil_progenitor3"  ,        
                                       "Eosinophil_progenitor4", 
                                       "Granulocyte_Monocyte_progenitor5",
                                       "Common_Myeloid_progenitor6")] %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(quantseq.modulescores,
          file = here("Rout/SA.HSC.Quantseq.modulescores.csv"))
```


```{r}
hvg.quantseq.modulescores = hvg.HSC.v2@meta.data[, c("Eosinophil1" , 
                                       "Neutrophil2" , 
                                       "Neutrophil_progenitor3"  ,        
                                       "Eosinophil_progenitor4", 
                                       "Granulocyte_Monocyte_progenitor5",
                                       "Common_Myeloid_progenitor6")] %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(hvg.quantseq.modulescores,
          file = here("Rout/SA.hvg.HSC.Quantseq.modulescores.csv"))
```



### Writing the modulescores for cytopus gene sets

```{r}
cytopus.modulescores = SA.HSC.modulescore.bm@meta.data %>% 
    select(abT.dat1:TSCM.dat53) %>% 
    rename_with(~str_remove(.x, ".dat[0-9]{1,2}$"))

write.csv(cytopus.modulescores,
          file = here("Rout/SA.HSC.cytopus.modulescores.csv"))
```


```{r}
hvg.cytopus.modulescores = hvg.HSC.v2@meta.data %>%
    select(abT.dat1:TSCM.dat52) %>% 
    rename_with(~str_remove(.x, ".dat[0-9]{1,2}$"))

write.csv(hvg.cytopus.modulescores,
          file = here("Rout/SA.hvg.HSC.cytopus.modulescores.csv"))
```

### Writing the modulescores for panglodb gene sets

```{r}
panglodb.modulescores = SA.HSC.modulescore.bm@meta.data %>% 
    select(abT.dat1:TSCM.dat53) %>% 
    rename_with(~str_remove(.x, ".dat[0-9]{1,2}$"))

write.csv(panglodb.modulescores,
          file = here("Rout/SA.HSC.panglodb.modulescores.csv"))
```


```{r}
hvg.panglodb.modulescores = hvg.HSC.v2@meta.data %>%
    select(B_cells_memory.txt1:T_regulatory_cells.txt21) %>% 
    rename_with(~str_remove(.x, ".txt[0-9]{1,2}$"))

write.csv(hvg.panglodb.modulescores,
          file = here("Rout/SA.hvg.HSC.panglodb.modulescores.csv"))
```


### Writing the modulescores for Jorssen et al 2024 gene sets

```{r}
table2.modulescores = SA.HSC.modulescore.bm@meta.data %>% 
    select(basophil1:neutrophil8) %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(table2.modulescores,
          file = here("Rout/SA.HSC.table2.modulescores.csv"))


table5.modulescores = SA.HSC.modulescore.bm@meta.data %>% 
    select(eos_stage11:eos_stage55) %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(table5.modulescores,
          file = here("Rout/SA.HSC.table5.modulescores.csv"))
```



```{r}
hvg.table2.modulescores = hvg.HSC.v2@meta.data %>% 
    select(basophil1:neutrophil8) %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(hvg.table2.modulescores,
          file = here("Rout/SA.hvg.HSC.table2.modulescores.csv"))

hvg.table5.modulescores = hvg.HSC.v2@meta.data %>% 
    select(eos_stage11:eos_stage55) %>% 
    rename_with(~str_remove(.x, "[0-9]$"))

write.csv(hvg.table5.modulescores,
          file = here("Rout/SA.hvg.HSC.table5.modulescores.csv"))
```
